UpdateScript/requirements.txt
psycopg2-binary==2.8.6
requests==2.28.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3